requirements.txt
simplegmail
pytest-playwright
# selenium==4.38.0
openpyxl
reportlab
